refactor(javm-cap): hoist Instance memory-overlay derivation to one place
The `(mem_size, overlays)` derivation from an Image's memory_mappings +
pinned/initial slot contents was copy-pasted in four spots — the recomp
bench harness, the kernel genesis path, and the two guest-test paths
(incl. the conformance oracle). The recompiler runtime and the
interpreter conformance oracle MUST agree on Instance memory layout, yet
each kept its own copy that could silently drift.

Replace all four with a single `Image::data_overlays()` method in
javm-cap, the canonical source of truth.

Smoke 12/12; conformance + sub_vm tests pass.
